+++ CHANGELOG.md
@@ -3,6 +3,11 @@ Changelog
 
 ## TBD
 
+### Enhancements
+
+* Added AppSession Span support for aggregating CPU and memory usage across manual app sessions, including long-running sessions and OS-supported background execution.
+  [#579](https://github.com/bugsnag/bugsnag-cocoa-performance/pull/579)
+
 ### Bug fixes
 
 * Support Pascal case plist configuration values.